Pull out shared helpers from MethodTableTests
